+++ CHANGELOG.md
@@ -47,6 +47,7 @@
 - Add layouts to contenttypes @robgietema
 - Add behaviors to fieldsets and properties in schema @robgietema
 - Add @export documentation and tests @robgietema
+- Add SearchableText metadata to the catalog @robgietema
 
 ### Bugfix